+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## 1.0.1 
+
+- Adds `APISchemaObject.isFreeForm` and related support for free-form schema objects.
+
+## 1.0.0
+
+- Adds support for OpenAPI 3.0
+- Splits Swagger (2.0) and OpenAPI (3.0) into 'package:open_api/v2.dart' and 'package:open_api/v3.dart'.
+
 ## 0.9.1
 
 - Moves shared properties from `APISchema`, `APIHeader` and `APIParameter` into `APIProperty`.